Respond to bookmark events by saving bookmarks
This adds the code that conditionally publishes bookmark events to
which applications can respond. The provided fragment responds by
popping a small icon in the bottom left corner of the screen, and
the demo application saves a bookmark.

Affects: https://www.notion.so/lyrasis/Audiobook-player-position-is-not-retained-after-exiting-the-app-602cd9ea0fe240feb2b02c338b154d1b