Explicit bookmarks (#1)
* Start implementing explicit bookmarks

This introduces a new event type that indicates that the hosting
application should save a bookmark. No code publishes this event yet.

Additionally, the demo application was imported into this project in
order to ease development.

* Upgrade libraries and dependencies for the demo app

* Respond to bookmark events by saving bookmarks

This adds the code that conditionally publishes bookmark events to
which applications can respond. The provided fragment responds by
popping a small icon in the bottom left corner of the screen, and
the demo application saves a bookmark.

Affects: https://www.notion.so/lyrasis/Audiobook-player-position-is-not-retained-after-exiting-the-app-602cd9ea0fe240feb2b02c338b154d1b

* Change the groupId and git repository names

* Add the infrastructure to test Findaway books

* Update presets