Brutal refactoring for PP-1074
The players required a ton of refactoring in order to properly handle
the new TOC structures. This makes the initial changes to drastically
simplify state handling and responsiveness of the open access player,
and to reduce the amount of duplication between the open access and
LCP player.

Currently, bookmarking is disabled, and the LCP player is effectively
disabled.

Affects: https://ebce-lyrasis.atlassian.net/browse/PP-1074