Remove opMovePlayheadToLocation optimization
The opMovePlayheadToLocation operation is used to position the playhead
when the book is first loaded. The problem is that this can/will fail:

  1. The current chapter of the book is set to the first chapter until something
     explicitly positions the playhead.
  2. The opMovePlayheadToLocation checks to see if the target position is the
     current chapter, in order to do an optimal seek() if it is.
  3. If the target location is the first chapter, and the book has just been
     opened, the above check will conclude that seek() is enough to set
     the position.
  4. The book has just been opened, therefore the chapter hasn't been loaded
     into ExoPlayer, therefore the seek will do nothing.

An examination of the code seems to indicate that we never depend on
opMovePlayheadToLocation being "optimal" anywhere; when the code wants to
explicitly seek (using the UI scrubber, for example), then it explicitly
seeks and doesn't use opMovePlayheadToLocation.

Affects: https://ebce-lyrasis.atlassian.net/browse/PP-1247