Change sleep timer to TOC item semantics
This makes the player publish "chapter completed" messages when it
is both playing, and the current TOC item is different to the one
that was current on the previous playback update.  The sleep timer
is called when chapters are completed (if it is configured to expire
on chapter changes).

Affects: https://ebce-lyrasis.atlassian.net/browse/PP-1267